pgs-50: Moved task list related views to new TaskListController and updated HomeController to improve url rendering.
