+++ Pegasus/Pegasus.csproj
@@ -21,4 +21,8 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Models\TaskList\" />
+  </ItemGroup>
+
 </Project>
