+++ Pegasus/Views/Shared/_Layout.cshtml
@@ -23,7 +23,7 @@
         <div class="container">
             <a asp-area="" asp-controller="Home" asp-action="Index" class="navbar-brand">Pegasus</a>
             <div class="nav-item float-right">
-                <a id="createButton" class="btn btn-primary" asp-controller="Home" asp-action="Create">Create Task</a>
+                <a id="createButton" class="btn btn-primary" asp-controller="TaskList" asp-action="Create">Create Task</a>
             </div>
             <button class="navbar-toggler" type="button" data-toggle="collapse" data-target=".navbar-collapse" aria-controls="navbarSupportedContent"
                     aria-expanded="false" aria-label="Toggle navigation">
@@ -33,7 +33,7 @@
             <div class="navbar-collapse collapse d-sm-inline-flex flex-sm-row-reverse">
                 <ul class="navbar-nav flex-grow-1">
                     <li class="nav-item">
-                        <a class="nav-link" asp-area="" asp-controller="Home" asp-action="Index">Task List</a>
+                        <a class="nav-link" asp-area="" asp-controller="TaskList" asp-action="Index">Task List</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link" asp-area="" asp-controller="Projects" asp-action="Index">Projects</a>
