+++ Pegasus/Views/TaskList/Create.cshtml
@@ -1,4 +1,4 @@
-@model Pegasus.Models.Home.TaskViewModel
+@model Pegasus.Models.TaskList.TaskViewModel
 
 @{
     ViewData["Title"] = "Task";
