+++ Pegasus/Views/TaskList/_EditTaskContent.cshtml
@@ -1,4 +1,4 @@
-@model Pegasus.Models.Home.TaskViewModel
+@model Pegasus.Models.TaskList.TaskViewModel
 
 <div id="editTaskContent" class="row">
     <div asp-validation-summary="ModelOnly" class="text-danger"></div>
